+++ src/components/ModalPopup.js
@@ -1,0 +1,31 @@
+import { Button, Modal, Tab, Tabs } from "react-bootstrap";
+
+export const ModalPopup = (props) => {
+    return (
+        <Modal show={props.show}>
+            <Modal.Header closeButton onClick={props.onHandleClick}>
+                <Modal.Title>{props.title}</Modal.Title>
+            </Modal.Header>
+            <Modal.Body>
+                <Tabs defaultActiveKey="payment">
+                    {
+                        props.tabs && props.tabs.map((item, index) => {
+                            return (
+                                <Tab eventKey={item.tabEventKey} title={item.tabTitle} key={index}>
+                                    {item.children}
+                                </Tab>
+                            )
+                        })
+                    }
+                </Tabs>
+            </Modal.Body>
+            <Modal.Footer>
+                <Button variant="secondary" onClick={props.onHandleClick}>
+                    Close
+                </Button>
+            </Modal.Footer>
+        </Modal>
+    );
+}
+
+export default ModalPopup;