+++ src/views/BankAccountTransactionsPage/BankAccountTransactionsPage.js
@@ -5,6 +5,12 @@ import Title from "../../components/Title";
 import { actions as bankAccountTranscationsActions } from "../../actions/app/BankAccountTransactionsActions";
 import DataTable from "react-data-table-component";
 import Amount from "../../components/Amount";
+import Date from "../../components/Date";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faEuroSign } from "@fortawesome/free-solid-svg-icons";
+import Toolbar from "../../components/Toolbar";
+import ModalPopup from "../../components/ModalPopup";
+import PaymentsForm from "../PaymentsPage/PaymentsForm";
 
 export const BankAccountTransactionsPage = (props) => {
 
@@ -12,10 +18,12 @@ export const BankAccountTransactionsPage = (props) => {
     const [perPageNumber, setPerPageNumber] = useState(10);
     const [orderBy, setOrderBy] = useState("transactionDate");
     const [orderWay, setOrderWay] = useState("desc");
+    const [show, setShow] = useState(false);
 
     const { items, index, total, isLoading } = props;
 
     useEffect(() => {
+        document.title = `${process.env.REACT_APP_NAME} - Bank Account Transactions`;
         index(pageNumber, perPageNumber, orderBy, orderWay);
     }, [index, pageNumber, perPageNumber, orderBy, orderWay]);
 
@@ -36,7 +44,7 @@ export const BankAccountTransactionsPage = (props) => {
             name: "Date",
             sortable: true,
             sortField: "transactionDate",
-            selector: row => row.transactionDate
+            selector: row => <Date date={row.transactionDate} />
         },
     ];
 
@@ -53,11 +61,45 @@ export const BankAccountTransactionsPage = (props) => {
         setOrderWay(sortDirection);
     }
 
+    const getToolbar = () => {
+        return [
+            {
+                handleProp: openModalPopup,
+                value: null,
+                variant: "success",
+                label: "make a payment",
+                icon: <FontAwesomeIcon icon={faEuroSign} />
+            }
+        ];
+    }
+
+    const openModalPopup = () => {
+        setShow(true);
+    }
+
+    const handleClick = () => {
+        setShow(false);
+    }
+    const buildTabsForModal = () => {
+        return [
+            {
+                tabEventKey: "payment",
+                tabTitle: "make a payment",
+                children: <PaymentsForm />,
+            },
+        ];
+    }
+
     return (
         <Container fluid>
             <Row>
                 <Col>
                     <Title label="Bank Account Transactions" />
+                </Col>
+            </Row>
+            <Row>
+                <Col>
+                    <Toolbar data={getToolbar()} />
                 </Col>
             </Row>
             <Row>
@@ -76,6 +118,12 @@ export const BankAccountTransactionsPage = (props) => {
                     />
                 </Col>
             </Row>
+            <ModalPopup
+                show={show}
+                onHandleClick={handleClick}
+                tabs={buildTabsForModal()}
+            >
+            </ModalPopup>
         </Container>
     )
 }